Update CI workflows to use actions/cache v5 and refactor search command argument parsing
This commit updates the CI workflows to utilize actions/cache version 5, ensuring compatibility with Node 24 runtime. Additionally, it refactors the argument parsing logic in the search command for improved clarity and maintainability, enhancing the overall user experience. These changes aim to streamline the development process and improve testing reliability.
