+++ .github/actions/setup-nimbus-ci/action.yml
@@ -24,7 +24,7 @@ runs:
         bun-version: ${{ inputs.bun-version }}
 
     - name: Cache Bun package download cache
-      uses: actions/cache@0057852bfaa89a56745cba8c7296529d2fc39830 # v4
+      uses: actions/cache@668228422ae6a00e4ad889ee87cd7109ec5666a7 # v5 (Node 24 runtime)
       with:
         path: ~/.bun/install/cache
         key: ${{ runner.os }}-bun-${{ hashFiles('bun.lock') }}
