+++ .github/workflows/ci.yml
@@ -138,7 +138,7 @@ jobs:
           run-checks: "false"
 
       - name: Cache Playwright browsers
-        uses: actions/cache@0057852bfaa89a56745cba8c7296529d2fc39830 # v4
+        uses: actions/cache@668228422ae6a00e4ad889ee87cd7109ec5666a7 # v5 (Node 24 runtime)
         with:
           path: .playwright-browsers
           key: ${{ runner.os }}-playwright-${{ hashFiles('bun.lock') }}
@@ -198,7 +198,7 @@ jobs:
           run-checks: "false"
 
       - name: Cache Playwright browsers
-        uses: actions/cache@0057852bfaa89a56745cba8c7296529d2fc39830 # v4
+        uses: actions/cache@668228422ae6a00e4ad889ee87cd7109ec5666a7 # v5 (Node 24 runtime)
         with:
           path: .playwright-browsers
           key: ${{ runner.os }}-playwright-${{ hashFiles('bun.lock') }}
