+++ .github/workflows/security.yml
@@ -53,6 +53,8 @@ jobs:
           format: "sarif"
           output: "trivy-results.sarif"
           ignore-unfixed: true
+          # v0.35.0 still pins actions/cache@v4 (Node 20); disable until trivy-action upgrades to cache v5.
+          cache: false
 
       - name: Upload Trivy results to GitHub Security tab
         uses: github/codeql-action/upload-sarif@c10b8064de6f491fea524254123dbe5e09572f13 # v4
@@ -100,7 +102,7 @@ jobs:
 
       - name: Cache cargo-audit binary
         id: cargo-audit-cache
-        uses: actions/cache@0057852bfaa89a56745cba8c7296529d2fc39830 # v4
+        uses: actions/cache@668228422ae6a00e4ad889ee87cd7109ec5666a7 # v5 (Node 24 runtime)
         with:
           path: ~/.cargo/bin/cargo-audit
           key: cargo-audit-0.22.1
