+++ packages/gateway/src/platform/assemble-sync-registrations.ts
@@ -1,0 +1,99 @@
+import { createBitbucketSyncable } from "../connectors/bitbucket-sync.ts";
+import { createConfluenceSyncable } from "../connectors/confluence-sync.ts";
+import { createDiscordSyncable } from "../connectors/discord-sync.ts";
+import { createGithubSyncable } from "../connectors/github-sync.ts";
+import { createGitlabSyncable } from "../connectors/gitlab-sync.ts";
+import { createGmailSyncable } from "../connectors/gmail-sync.ts";
+import { createGoogleDriveSyncable } from "../connectors/google-drive-sync.ts";
+import { createGooglePhotosSyncable } from "../connectors/google-photos-sync.ts";
+import { createJiraSyncable } from "../connectors/jira-sync.ts";
+import type { LazyConnectorMesh } from "../connectors/lazy-mesh.ts";
+import { createLinearSyncable } from "../connectors/linear-sync.ts";
+import { createNotionSyncable } from "../connectors/notion-sync.ts";
+import { createOneDriveSyncable } from "../connectors/onedrive-sync.ts";
+import { createOutlookSyncable } from "../connectors/outlook-sync.ts";
+import { createSlackSyncable } from "../connectors/slack-sync.ts";
+import { createTeamsSyncable } from "../connectors/teams-sync.ts";
+import type { SyncScheduler } from "../sync/scheduler.ts";
+
+/** Registers all connector-backed sync jobs that lazily ensure MCP children via the mesh. */
+export function registerConnectorMeshSyncables(
+  syncScheduler: SyncScheduler,
+  connectorMesh: LazyConnectorMesh,
+): void {
+  syncScheduler.register(
+    createGoogleDriveSyncable({
+      ensureGoogleDriveRunning: () => connectorMesh.ensureGoogleDriveRunning(),
+    }),
+  );
+  syncScheduler.register(
+    createGmailSyncable({
+      ensureGoogleMcpRunning: () => connectorMesh.ensureGoogleDriveRunning(),
+    }),
+  );
+  syncScheduler.register(
+    createGooglePhotosSyncable({
+      ensureGoogleMcpRunning: () => connectorMesh.ensureGoogleDriveRunning(),
+    }),
+  );
+  syncScheduler.register(
+    createOneDriveSyncable({
+      ensureMicrosoftMcpRunning: () => connectorMesh.ensureMicrosoftBundleRunning(),
+    }),
+  );
+  syncScheduler.register(
+    createOutlookSyncable({
+      ensureMicrosoftMcpRunning: () => connectorMesh.ensureMicrosoftBundleRunning(),
+    }),
+  );
+  syncScheduler.register(
+    createGithubSyncable({
+      ensureGithubMcpRunning: () => connectorMesh.ensureGithubRunning(),
+    }),
+  );
+  syncScheduler.register(
+    createGitlabSyncable({
+      ensureGitlabMcpRunning: () => connectorMesh.ensureGitlabRunning(),
+    }),
+  );
+  syncScheduler.register(
+    createBitbucketSyncable({
+      ensureBitbucketMcpRunning: () => connectorMesh.ensureBitbucketRunning(),
+    }),
+  );
+  syncScheduler.register(
+    createSlackSyncable({
+      ensureSlackMcpRunning: () => connectorMesh.ensureSlackRunning(),
+    }),
+  );
+  syncScheduler.register(
+    createTeamsSyncable({
+      ensureMicrosoftMcpRunning: () => connectorMesh.ensureMicrosoftBundleRunning(),
+    }),
+  );
+  syncScheduler.register(
+    createLinearSyncable({
+      ensureLinearMcpRunning: () => connectorMesh.ensureLinearRunning(),
+    }),
+  );
+  syncScheduler.register(
+    createJiraSyncable({
+      ensureJiraMcpRunning: () => connectorMesh.ensureJiraRunning(),
+    }),
+  );
+  syncScheduler.register(
+    createNotionSyncable({
+      ensureNotionMcpRunning: () => connectorMesh.ensureNotionRunning(),
+    }),
+  );
+  syncScheduler.register(
+    createConfluenceSyncable({
+      ensureConfluenceMcpRunning: () => connectorMesh.ensureConfluenceRunning(),
+    }),
+  );
+  syncScheduler.register(
+    createDiscordSyncable({
+      ensureDiscordMcpRunning: () => connectorMesh.ensureDiscordRunning(),
+    }),
+  );
+}
