+++ packages/gateway/src/platform/assemble.ts
@@ -3,29 +3,20 @@ import { join } from "node:path";
 import pino from "pino";
 import { loadNimbusEmbeddingFromConfigDir } from "../config/nimbus-toml.ts";
 import { Config } from "../config.ts";
-import { createBitbucketSyncable } from "../connectors/bitbucket-sync.ts";
-import { createConfluenceSyncable } from "../connectors/confluence-sync.ts";
-import { createDiscordSyncable } from "../connectors/discord-sync.ts";
-import { createGithubSyncable } from "../connectors/github-sync.ts";
-import { createGitlabSyncable } from "../connectors/gitlab-sync.ts";
-import { createGmailSyncable } from "../connectors/gmail-sync.ts";
-import { createGoogleDriveSyncable } from "../connectors/google-drive-sync.ts";
-import { createGooglePhotosSyncable } from "../connectors/google-photos-sync.ts";
-import { createJiraSyncable } from "../connectors/jira-sync.ts";
 import { createLazyConnectorMesh } from "../connectors/lazy-mesh.ts";
-import { createLinearSyncable } from "../connectors/linear-sync.ts";
-import { createNotionSyncable } from "../connectors/notion-sync.ts";
-import { createOneDriveSyncable } from "../connectors/onedrive-sync.ts";
-import { createOutlookSyncable } from "../connectors/outlook-sync.ts";
-import { createSlackSyncable } from "../connectors/slack-sync.ts";
-import { createTeamsSyncable } from "../connectors/teams-sync.ts";
 import { createEmbeddingRuntime } from "../embedding/create-embedding-runtime.ts";
 import { LOCAL_EMBEDDING_MODEL_ID } from "../embedding/model.ts";
-import { LocalIndex } from "../index/local-index.ts";
+import {
+  LocalIndex,
+  type LocalIndexOptions,
+  type SemanticSearchDeps,
+} from "../index/local-index.ts";
 import { createIpcServer } from "../ipc/index.ts";
 import { ProviderRateLimiter } from "../sync/rate-limiter.ts";
 import { SyncScheduler } from "../sync/scheduler.ts";
+import type { SyncContext } from "../sync/types.ts";
 import { createNimbusVault } from "../vault/factory.ts";
+import { registerConnectorMeshSyncables } from "./assemble-sync-registrations.ts";
 import { openUrlInDefaultBrowser } from "./browser.ts";
 import { ensurePlatformDirectories } from "./dirs.ts";
 import { processEnvGet } from "./env-access.ts";
@@ -66,126 +57,61 @@ export async function assemblePlatformServices(paths: PlatformPaths): Promise<Pl
     Config.embeddingsEnabled,
   );
   const rt = embeddingRuntime;
-  const scheduleItemEmbedding = rt?.scheduleItemEmbedding.bind(rt);
-  const semanticSearch =
-    rt != null
-      ? {
-          model: LOCAL_EMBEDDING_MODEL_ID,
-          embedQuery: (text: string) => rt.embedQuery(text),
-        }
-      : undefined;
-  const localOpts =
-    scheduleItemEmbedding !== undefined || semanticSearch !== undefined
-      ? {
-          ...(scheduleItemEmbedding !== undefined ? { scheduleItemEmbedding } : {}),
-          ...(semanticSearch !== undefined ? { semanticSearch } : {}),
-        }
-      : undefined;
-  const localIndex = localOpts !== undefined ? new LocalIndex(db, localOpts) : new LocalIndex(db);
-  const syncContext =
-    scheduleItemEmbedding !== undefined
-      ? { vault, db, logger: syncLogger, rateLimiter, scheduleItemEmbedding }
-      : { vault, db, logger: syncLogger, rateLimiter };
+  let scheduleItemEmbedding: ((itemId: string) => void) | undefined;
+  let semanticSearch: SemanticSearchDeps | undefined;
+  if (rt) {
+    scheduleItemEmbedding = rt.scheduleItemEmbedding.bind(rt);
+    semanticSearch = {
+      model: LOCAL_EMBEDDING_MODEL_ID,
+      embedQuery: (text: string) => rt.embedQuery(text),
+    };
+  }
+  const localIndexOpts: LocalIndexOptions = {};
+  if (scheduleItemEmbedding !== undefined) {
+    localIndexOpts.scheduleItemEmbedding = scheduleItemEmbedding;
+  }
+  if (semanticSearch !== undefined) {
+    localIndexOpts.semanticSearch = semanticSearch;
+  }
+  const hasEmbeddingIndexOpts = scheduleItemEmbedding !== undefined || semanticSearch !== undefined;
+  let localIndex: LocalIndex;
+  if (hasEmbeddingIndexOpts) {
+    localIndex = new LocalIndex(db, localIndexOpts);
+  } else {
+    localIndex = new LocalIndex(db);
+  }
+  const syncBase: SyncContext = { vault, db, logger: syncLogger, rateLimiter };
+  let syncContext: SyncContext;
+  if (scheduleItemEmbedding !== undefined) {
+    syncContext = { ...syncBase, scheduleItemEmbedding };
+  } else {
+    syncContext = syncBase;
+  }
   const syncScheduler = new SyncScheduler(syncContext, undefined, {
     notify: async (title, body) => {
       await notifications.show(title, body);
     },
   });
   const connectorMesh = await createLazyConnectorMesh(paths, vault);
-  syncScheduler.register(
-    createGoogleDriveSyncable({
-      ensureGoogleDriveRunning: () => connectorMesh.ensureGoogleDriveRunning(),
-    }),
-  );
-  syncScheduler.register(
-    createGmailSyncable({
-      ensureGoogleMcpRunning: () => connectorMesh.ensureGoogleDriveRunning(),
-    }),
-  );
-  syncScheduler.register(
-    createGooglePhotosSyncable({
-      ensureGoogleMcpRunning: () => connectorMesh.ensureGoogleDriveRunning(),
-    }),
-  );
-  syncScheduler.register(
-    createOneDriveSyncable({
-      ensureMicrosoftMcpRunning: () => connectorMesh.ensureMicrosoftBundleRunning(),
-    }),
-  );
-  syncScheduler.register(
-    createOutlookSyncable({
-      ensureMicrosoftMcpRunning: () => connectorMesh.ensureMicrosoftBundleRunning(),
-    }),
-  );
-  syncScheduler.register(
-    createGithubSyncable({
-      ensureGithubMcpRunning: () => connectorMesh.ensureGithubRunning(),
-    }),
-  );
-  syncScheduler.register(
-    createGitlabSyncable({
-      ensureGitlabMcpRunning: () => connectorMesh.ensureGitlabRunning(),
-    }),
-  );
-  syncScheduler.register(
-    createBitbucketSyncable({
-      ensureBitbucketMcpRunning: () => connectorMesh.ensureBitbucketRunning(),
-    }),
-  );
-  syncScheduler.register(
-    createSlackSyncable({
-      ensureSlackMcpRunning: () => connectorMesh.ensureSlackRunning(),
-    }),
-  );
-  syncScheduler.register(
-    createTeamsSyncable({
-      ensureMicrosoftMcpRunning: () => connectorMesh.ensureMicrosoftBundleRunning(),
-    }),
-  );
-  syncScheduler.register(
-    createLinearSyncable({
-      ensureLinearMcpRunning: () => connectorMesh.ensureLinearRunning(),
-    }),
-  );
-  syncScheduler.register(
-    createJiraSyncable({
-      ensureJiraMcpRunning: () => connectorMesh.ensureJiraRunning(),
-    }),
-  );
-  syncScheduler.register(
-    createNotionSyncable({
-      ensureNotionMcpRunning: () => connectorMesh.ensureNotionRunning(),
-    }),
-  );
-  syncScheduler.register(
-    createConfluenceSyncable({
-      ensureConfluenceMcpRunning: () => connectorMesh.ensureConfluenceRunning(),
-    }),
-  );
-  syncScheduler.register(
-    createDiscordSyncable({
-      ensureDiscordMcpRunning: () => connectorMesh.ensureDiscordRunning(),
-    }),
-  );
+  registerConnectorMeshSyncables(syncScheduler, connectorMesh);
   syncScheduler.start();
   rt?.startBackgroundJobs();
+  const ipcOpts: Parameters<typeof createIpcServer>[0] = {
+    listenPath: paths.socketPath,
+    vault,
+    version: "0.1.0",
+    localIndex,
+    openUrl: openUrlInDefaultBrowser,
+    syncScheduler,
+  };
+  if (rt) {
+    ipcOpts.getEmbeddingStatus = () => ({
+      embeddingBackfill: rt.getBackfillProgress(),
+    });
+  }
   return {
     vault,
-    ipc: createIpcServer({
-      listenPath: paths.socketPath,
-      vault,
-      version: "0.1.0",
-      localIndex,
-      openUrl: openUrlInDefaultBrowser,
-      syncScheduler,
-      ...(rt != null
-        ? {
-            getEmbeddingStatus: () => ({
-              embeddingBackfill: rt.getBackfillProgress(),
-            }),
-          }
-        : {}),
-    }),
+    ipc: createIpcServer(ipcOpts),
     paths,
     localIndex,
     connectorMesh,
